requirements-backend.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27
aioimaplib>=1.0.1
aiosmtplib>=3.0
alembic>=1.13
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20
psycopg[binary]>=3.2
email-validator>=2.0
bleach>=6.0
cryptography>=42
google-api-python-client>=2.120.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
structlog>=24.1
python-json-logger>=2.0
prometheus-fastapi-instrumentator>=7.0
opentelemetry-api>=1.27
opentelemetry-sdk>=1.27
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-sqlalchemy>=0.48b0
opentelemetry-exporter-otlp>=1.27
sentry-sdk[fastapi]>=2.18
